chore(outlier): ♻️ move function to a new internal.outlier module
This should indicate that the function is not a part of the public API. Thanks, Jonas Mueller <1390638+jwmueller@users.noreply.github.com>!
