+++ docs/source/cleanlab/internal/index.rst
@@ -17,5 +17,6 @@ internal
     label_quality_utils
     multilabel_utils
     multilabel_scorer
+    outlier
     token_classification_utils
     validation
