+++ docs/source/cleanlab/internal/outlier.rst
@@ -1,0 +1,8 @@
+outlier
+=======
+
+.. automodule:: cleanlab.internal.outlier
+    :autosummary:
+    :members:
+    :undoc-members:
+    :show-inheritance:
